fix: inherit semantic commit type from internal dependency updates
By default, all dependency updates are committed with "chore(deps)"
which, by default, does not cause a version bump with semantic-release.
This change causes internal dependency updates to be treated differently
such that the semantic commit type is inherited. For example, if
scratch-vm updates from 3.1.4 to 3.2.0, a minor bump indicating a
feature update, then the corresponding PR in scratch-gui will be marked
"feat" to cause a corresponding minor bump in scratch-gui.

Without this change, updates to internal dependencies will be "merged up
the chain" as expected but they will not trigger a release.